HePPCAT: HEteroscedastic Probabilistic PCA Technique

Documentation for HePPCAT.

👋 This package provides research code and work is ongoing. If you are interested in using it in your own research, I'd love to hear from you and collaborate! Feel free to write: dahong67@wharton.upenn.edu

Please cite the following paper for this technique:

David Hong, Kyle Gilman, Laura Balzano, Jeffrey A. Fessler. "HePPCAT: Probabilistic PCA for Data with Heteroscedastic Noise", IEEE Transactions on Signal Processing 69:4819-4834, Aug. 2021. https://doi.org/10.1109/TSP.2021.3104979 https://arxiv.org/abs/2101.03468.

In BibTeX form:

@article{hgbf2021heppcat,
  title   = "{HePPCAT}: Probabilistic {PCA} for Data with Heteroscedastic Noise",
  author  = "David Hong and Kyle Gilman and Laura Balzano and Jeffrey A. Fessler",
  journal = "{IEEE} Transactions on Signal Processing",
  year    = "2021",
  volume  = "69",
  pages   = "4819--4834",
  DOI     = "10.1109/tsp.2021.3104979",
}

Docstrings

HePPCAT.HePPCAT

HePPCAT.ArmijoSearch

HePPCAT.CubicSolvableMinorizer

HePPCAT.DifferenceOfConcave

HePPCAT.ExpectationMaximization

HePPCAT.HePPCATModel

HePPCAT.HePPCATModel

HePPCAT.HePPCATModel

HePPCAT.InverseLipschitz

HePPCAT.MinorizeMaximize

HePPCAT.OptimalQuadraticMinorizer

HePPCAT.ProjectedGradientAscent

HePPCAT.ProjectedVariance

HePPCAT.QuadraticMinorizer

HePPCAT.QuadraticSolvableMinorizer

HePPCAT.RootFinding

HePPCAT.StiefelGradientAscent

HePPCAT.F

HePPCAT.LipBoundU1

HePPCAT.LipBoundU2

HePPCAT.geodesic

HePPCAT.gradF

HePPCAT.heppcat

HePPCAT.homppca

HePPCAT.loglikelihood

HePPCAT.polar

HePPCAT.skew

HePPCAT.updateF!

HePPCAT.updateF!

HePPCAT.updateU!

HePPCAT.updateU!

HePPCAT.updateU!

HePPCAT.updateU!

HePPCAT.updatev!

HePPCAT.updatevl

HePPCAT.updatevl

HePPCAT.updatevl

HePPCAT.updatevl

HePPCAT.updatevl

HePPCAT.updatevl

HePPCAT.updatevl

HePPCAT.updateλ!

HePPCAT.updateλj

HePPCAT.updateλj

HePPCAT.updateλj

HePPCAT.updateλj

HePPCAT.updateλj

HePPCAT.updateλj

HePPCAT.updateλj

HePPCAT.HePPCAT — Module

Heteroscedastic PPCA Technique (HePPCAT) module. Provides probabilistic PCA for data with samples of heterogeneous quality (heteroscedastic noise).

HePPCAT.ArmijoSearch — Type

ArmijoSearch{S<:Integer,T<:AbstractFloat}

Armijo line search with parameters:

maxsearches :: S maximum number of line searches

stepsize :: T initial stepsize

contraction :: T contraction factor

tol :: T tolerance for sufficient decrease

HePPCAT.CubicSolvableMinorizer — Type

CubicSolvableMinorizer

Minorize maximize update with cubic solvable minorizer. No fields.

HePPCAT.DifferenceOfConcave — Type

DifferenceOfConcave

Difference of concave update method. No fields.

HePPCAT.ExpectationMaximization — Type

ExpectationMaximization

Expectation maximization update method. No fields.

HePPCAT.HePPCATModel — Type

HePPCATModel{S<:Number,T<:Real}

Model parameters for probabilistic PCA with noise that is heteroscedastic across samples. This is the return type of heppcat(_), the corresponding estimation function.

Properties

F :: Matrix factor matrix (computed via F = U*sqrt(Diagonal(λ))*Vt)

U :: Matrix{S} eigenvectors of factor covariance F*F'

λ :: Vector{T} eigenvalues of factor covariance F*F' (spike eigenvalues)

Vt :: Matrix{S} (transposed) eigenvectors of F'*F (i.e., right singular vectors of F)

v :: Vector{T} noise variances

HePPCAT.HePPCATModel — Method

HePPCATModel(U::AbstractMatrix,λ::AbstractVector,Vt::AbstractMatrix,v::AbstractVector)

Construct HePPCATModel object from factor eigenstructure and noise variances.

HePPCAT.HePPCATModel — Method

HePPCATModel(F::AbstractMatrix,v::AbstractVector)

Construct HePPCATModel object from factor matrix and noise variances.

HePPCAT.InverseLipschitz — Type

InverseLipschitz{T}

Inverse Lipschitz step size using bound function specified by field bound::T. Default choice is bound=LipBoundU2.

HePPCAT.MinorizeMaximize — Type

MinorizeMaximize

Minorize maximize update method. No fields.

HePPCAT.OptimalQuadraticMinorizer — Type

OptimalQuadraticMinorizer

Minorize maximize update using quadratic minorizer with optimized curvature. No fields.

HePPCAT.ProjectedGradientAscent — Type

ProjectedGradientAscent{T}

Projected gradient ascent update method. One field: stepsize::T.

HePPCAT.ProjectedVariance — Type

ProjectedVariance{S,T}

Update variance using method then project the variance so that the variance iterate does not go below varfloor.

HePPCAT.QuadraticMinorizer — Type

QuadraticMinorizer

Minorize maximize update using quadratic minorizer. No fields.

HePPCAT.QuadraticSolvableMinorizer — Type

QuadraticSolvableMinorizer

Minorize maximize update with quadratic solvable minorizer. No fields.

HePPCAT.RootFinding — Type

RootFinding

Root-finding (global maximization) update method. No fields.

HePPCAT.StiefelGradientAscent — Type

StiefelGradientAscent{T}

Stiefel gradient ascent update method. One field: stepsize::T.

HePPCAT.F — Method

F(U,λ,v,Y)

Objective for the optimization problem w.r.t U.

HePPCAT.LipBoundU1 — Method

LipBoundU1(M::HePPCATModel,Y)

Lipschitz bound of objective w.r.t. U at M with data Y.

HePPCAT.LipBoundU2 — Method

LipBoundU2(M::HePPCATModel,Y)

Lipschitz bound of objective w.r.t. U at M with data Y.

HePPCAT.geodesic — Method

geodesic(U,X,t)

Geodesic step of size t from U in direction of X along the Stiefel manifold.

HePPCAT.gradF — Method

gradF(U,λ,v,Y)

Euclidean gradient of objective for the optimization problem w.r.t U.

HePPCAT.heppcat — Method

heppcat(Y,k,iters;init=homppca(Y,k),vknown=false)

Estimate probabilistic PCA model for noise that is heteroscedastic across samples.

Inputs

Y : list of matrices (each column is a sample)

k : number of factors

iters : number of iterations to run

init : initial model (will be modified in-place)

vknown : variances are known (do not update) default false

varfloor : lower bound for variance iterates (useful if the iterates are degenerating to zero) default = 0

Output is a HePPCATModel object.

HePPCAT.homppca — Method

homppca(Y,k)

Estimate probabilistic PCA model for noise that is homoscedastic across samples.

Inputs are:

Y : list of matrices (each column is a sample)

k : number of factors

Output is a HePPCATModel object.

HePPCAT.loglikelihood — Method

loglikelihood(M::HePPCATModel,Y)

Log-likelihood of model M with respect to data Y (dropping constant term).

HePPCAT.polar — Method

polar(A)

Polar factor of A.

HePPCAT.skew — Method

skew(A)

Skew-symmetrize A by computing (A-A')/2.

HePPCAT.updateF! — Function

updateF!(M::HePPCATModel,Y,method)

Update the factor matrix in M with respect to data Y using method.

HePPCAT.updateF! — Method

updateF!(M::HePPCATModel,Y,::ExpectationMaximization)

Expectation maximization update of F.

HePPCAT.updateU! — Function

updateU!(M::HePPCATModel,Y,method)

Update the factor eigenvectors M.U with respect to data Y using method.

HePPCAT.updateU! — Method

updateU!(M::HePPCATModel,Y,::MinorizeMaximize)

Minorize maximize update of M.U using a linear minorizer.

HePPCAT.updateU! — Method

updateU!(M::HePPCATModel,Y,pga::ProjectedGradientAscent)

Projected gradient ascent update of M.U. Supported step size types: Number (i.e., constant step size), InverseLipschitz, and ArmijoSearch (experimental).

HePPCAT.updateU! — Method

updateU!(M::HePPCATModel,Y,sga::StiefelGradientAscent)

Stiefel gradient ascent update of M.U. Supported step size types: ArmijoSearch.

HePPCAT.updatev! — Method

updatev!(M::HePPCATModel,Y,method)

Update the noise variances M.v with respect to data Y using method. Internally calls updatevl(_) to update each entry.

HePPCAT.updatevl — Function

updatevl(vl,U,λ,Yl,method)

Compute update for lth noise variance vl with respect to data Yl using method.

HePPCAT.updatevl — Method

updatevl(vl,U,λ,Yl,::CubicSolvableMinorizer)

Minorize maximize update of vl with cubic solvable minorizer.

HePPCAT.updatevl — Method

updatevl(vl,U,λ,Yl,::DifferenceOfConcave)

Difference of concave update of vl.

HePPCAT.updatevl — Method

updatevl(vl,U,λ,Yl,::ExpectationMaximization)

Expectation maximization update of vl.

HePPCAT.updatevl — Method

updatevl(vl,U,λ,Yl,method::ProjectedVariance)

Update variance using method.method then project the variance so that the variance iterate does not go below method.varfloor.

HePPCAT.updatevl — Method

updatevl(vl,U,λ,Yl,::QuadraticSolvableMinorizer)

Minorize maximize update of vl with quadratic solvable minorizer.

HePPCAT.updatevl — Method

updatevl(vl,U,λ,Yl,::RootFinding)

Root-finding (global maximization) update of vl.

HePPCAT.updateλ! — Method

updateλ!(M::HePPCATModel,Y,method)

Update the factor eigenvalues M.λ with respect to data Y using method. Internally calls updateλj(_) to update each entry.

HePPCAT.updateλj — Function

updateλj(λj,uj,v,Y,method)

Compute update for jth factor eigenvalue λj with respect to data Y using method.

HePPCAT.updateλj — Method

updateλj(λj,uj,v,Y,::DifferenceOfConcave)

Difference of concave update of λj.

HePPCAT.updateλj — Method

updateλj(λj,uj,v,Y,::ExpectationMaximization)

Expectation maximization update of λj.

HePPCAT.updateλj — Method

updateλj(λj,uj,v,Y,::MinorizeMaximize)

Minorize maximize update of λj using the minorizer from

Y. Sun, A. Breloy, P. Babu, D. P. Palomar, F. Pascal, and G. Ginolhac, "Low-complexity algorithms for low rank clutter parameters estimation in radar systems," IEEE Transactions on Signal Processing, vol. 64, no. 8, pp. 1986–1998, Apr. 2016.

HePPCAT.updateλj — Method

updateλj(λj,uj,v,Y,::OptimalQuadraticMinorizer)

Minorize maximize update of λj using quadratic minorizer with optimized curvature.

HePPCAT.updateλj — Method

updateλj(λj,uj,v,Y,::QuadraticMinorizer)

Minorize maximize update of λj using quadratic minorizer.

HePPCAT.updateλj — Method

updateλj(λj,uj,v,Y,::RootFinding)

Root-finding (global maximization) update of λj.